requirements.txt
matplotlib==1.5.0
pyserial==2.7
numpy==1.10.1